refactor(welcome): extract shared WelcomeStepTemplate from language step
