+++ lib/features/welcome/language_step.dart
@@ -1,9 +1,9 @@
 import "dart:ui" as ui;
 
 import "package:eve_fit_assistant/config/locale.dart";
+import "package:eve_fit_assistant/features/welcome/welcome_step_template.dart";
 import "package:eve_fit_assistant/storage/setting/setting.dart";
 import "package:eve_fit_assistant/utils/context.dart";
-import "package:eve_fit_assistant/utils/screen.dart";
 import "package:flutter/material.dart" hide Locale;
 import "package:flutter_riverpod/flutter_riverpod.dart";
 
@@ -22,31 +22,30 @@ class LanguageStepPage extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final selected = ref.watch(localeProvider);
+    final detectedCode = ui.PlatformDispatcher.instance.locale.languageCode;
 
-    return Scaffold(
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
-          child: screenColumnTarget(context) == ScreenColumnTarget.one
-              ? _PhoneLayout(
-                  selected: selected,
-                  onSelected: (value) => ref
-                      .read(appSettingServiceProvider.notifier)
-                      .update((old) => old.copyWith(locale: value)),
-                  onContinue: onContinue,
-                  onSkip: onSkip,
-                  onBack: onBack,
-                )
-              : _TabletLayout(
-                  selected: selected,
-                  onSelected: (value) => ref
-                      .read(appSettingServiceProvider.notifier)
-                      .update((old) => old.copyWith(locale: value)),
-                  onContinue: onContinue,
-                  onSkip: onSkip,
-                  onBack: onBack,
-                ),
-        ),
+    void onSelected(Locale value) =>
+        ref.read(appSettingServiceProvider.notifier).update((old) => old.copyWith(locale: value));
+
+    return WelcomeStepTemplate(
+      title: context.l10n.welcomeLanguageTitle,
+      subtitle: context.l10n.welcomeLanguageSubtitle,
+      onContinue: onContinue,
+      onSkip: onSkip,
+      onBack: onBack,
+      content: Column(
+        children: [
+          for (final locale in Locale.values)
+            Padding(
+              padding: const EdgeInsets.only(bottom: 12),
+              child: _LanguageCard(
+                locale: locale,
+                isSelected: locale == selected,
+                isDetected: locale.name == detectedCode,
+                onTap: () => onSelected(locale),
+              ),
+            ),
+        ],
       ),
     );
   }
@@ -109,168 +108,6 @@ class _LanguageCard extends StatelessWidget {
               if (isSelected) Icon(Icons.check, color: colorScheme.primary, size: 22),
             ],
           ),
-        ),
-      ),
-    );
-  }
-}
-
-class _PhoneLayout extends StatelessWidget {
-  const _PhoneLayout({
-    required this.selected,
-    required this.onSelected,
-    required this.onContinue,
-    required this.onSkip,
-    required this.onBack,
-  });
-
-  final Locale selected;
-  final void Function(Locale) onSelected;
-  final VoidCallback onContinue;
-  final VoidCallback onSkip;
-  final VoidCallback onBack;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-    final detectedCode = ui.PlatformDispatcher.instance.locale.languageCode;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        const SizedBox(height: 32),
-        Text(context.l10n.welcomeLanguageTitle, style: theme.textTheme.headlineMedium),
-        const SizedBox(height: 4),
-        Text(
-          context.l10n.welcomeLanguageSubtitle,
-          style: theme.textTheme.bodyLarge?.copyWith(color: colorScheme.onSurfaceVariant),
-        ),
-        const SizedBox(height: 32),
-        for (final locale in Locale.values)
-          Padding(
-            padding: const EdgeInsets.only(bottom: 12),
-            child: _LanguageCard(
-              locale: locale,
-              isSelected: locale == selected,
-              isDetected: locale.name == detectedCode,
-              onTap: () => onSelected(locale),
-            ),
-          ),
-        const SizedBox(height: 24),
-        Center(
-          child: FilledButton(
-            onPressed: onContinue,
-            child: Text(context.l10n.welcomeContinueButton),
-          ),
-        ),
-        const SizedBox(height: 16),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: [
-            TextButton(onPressed: onBack, child: Text(context.l10n.welcomeBackButton)),
-            TextButton(onPressed: onSkip, child: Text(context.l10n.welcomeSkipButton)),
-          ],
-        ),
-      ],
-    );
-  }
-}
-
-class _TabletLayout extends StatelessWidget {
-  const _TabletLayout({
-    required this.selected,
-    required this.onSelected,
-    required this.onContinue,
-    required this.onSkip,
-    required this.onBack,
-  });
-
-  final Locale selected;
-  final void Function(Locale) onSelected;
-  final VoidCallback onContinue;
-  final VoidCallback onSkip;
-  final VoidCallback onBack;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-    final detectedCode = ui.PlatformDispatcher.instance.locale.languageCode;
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 80, vertical: 32),
-      child: Center(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            Row(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.end,
-                    children: [
-                      Text(
-                        context.l10n.welcomeLanguageTitle,
-                        style: theme.textTheme.headlineMedium,
-                        textAlign: TextAlign.right,
-                      ),
-                      const SizedBox(height: 8),
-                      Text(
-                        context.l10n.welcomeLanguageSubtitle,
-                        style: theme.textTheme.bodyLarge?.copyWith(
-                          color: colorScheme.onSurfaceVariant,
-                        ),
-                        textAlign: TextAlign.right,
-                      ),
-                    ],
-                  ),
-                ),
-                Expanded(
-                  flex: 2,
-                  child: Padding(
-                    padding: const EdgeInsets.only(left: 48),
-                    child: Column(
-                      children: [
-                        for (final locale in Locale.values)
-                          Padding(
-                            padding: const EdgeInsets.only(bottom: 16),
-                            child: _LanguageCard(
-                              locale: locale,
-                              isSelected: locale == selected,
-                              isDetected: locale.name == detectedCode,
-                              onTap: () => onSelected(locale),
-                            ),
-                          ),
-                      ],
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 40),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 4),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Row(
-                    children: [
-                      TextButton(onPressed: onBack, child: Text(context.l10n.welcomeBackButton)),
-                      const SizedBox(width: 8),
-                      TextButton(onPressed: onSkip, child: Text(context.l10n.welcomeSkipButton)),
-                    ],
-                  ),
-                  FilledButton(
-                    onPressed: onContinue,
-                    child: Text(context.l10n.welcomeContinueButton),
-                  ),
-                ],
-              ),
-            ),
-          ],
         ),
       ),
     );
